+++ programs/cp-amm/src/tests/test_operator_permission.rs
@@ -2,7 +2,7 @@ use crate::state::{Operator, OperatorPermission};
 
 #[test]
 fn test_initialize_with_full_permission() {
-    let permission: u128 = 0b111111111111;
+    let permission: u128 = 0b1111111111111;
     assert!(
         permission > 1 << (OperatorPermission::VARIANT_COUNT - 1)
             && permission < 1 << OperatorPermission::VARIANT_COUNT
@@ -59,6 +59,10 @@ fn test_initialize_with_full_permission() {
     );
     assert_eq!(
         operator.is_permission_allow(OperatorPermission::FixPool),
+        true
+    );
+    assert_eq!(
+        operator.is_permission_allow(OperatorPermission::UpdateConfigPermission),
         true
     );
 }